pyproject.toml
[tool.poetry]
name = "IntervalRepetition"
version = "0.1.0"
description = ""
authors = ["TamaraMalyheva <tamaramalysheva5991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.4"
pymongo = "^4.6.1"
python-dotenv = "^1.0.1"
aiogram = "^3.4.1"
apscheduler = "^3.10.4"
pytest = "^8.0.2"
pytest-mock = "^3.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
